A lot of things happened, shortlist: rewired many Main.grid dependencies, added proper wavestart button, made temple, which contains floors, squads and projectiles, wired everything through it, cleaned up a lot of code, and more.

+++ src/explorerTypes/Exploder.java
@@ -33,7 +33,7 @@ public class Exploder extends Explorer{
 		
 	}
 	@Override 
-	public void tick(int milli){
+	public void tick(long milli){
 		cd-=milli;
 	}
